server/db/models/Brand.js: created file and added contents, server/db/index.js: added Product & Brand models, added Product/Brand associations, server/db/models/Product.js: migrated removed associations/Brand model code
